requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
numpy
Pillow
gunicorn
eventlet==0.33.3